mongo

+++ code/base_db.py
@@ -10,8 +10,8 @@ from bson.decimal128 import Decimal128
 class DecimalEncoder(json.JSONEncoder):
     def default(self, obj):
         if isinstance(obj, Decimal):
-            return Decimal128(str(obj))
-            
+            return float(str(obj))
+
         return json.JSONEncoder.default(self, obj)
